feat: 2025 updates
Begin making some updates. Start with updating all the requirement versions.
Secondly, change the way we do testing. Remove all the old test data and we'll rely on mocks entirely.
Set all failing tests because of this to 'xfail' and we'll fix them as we go.

+++ discordbot/betmanager.py
@@ -107,7 +107,7 @@ class BetManager:
 
         # add on loser points
         with contextlib.suppress(ValueError, ZeroDivisionError, TypeError):
-            points_won += int(math.floor(extra_eddies / num_winners))
+            points_won += math.floor(extra_eddies / num_winners)
         return points_won
 
     def _calculate_taxed_winnings(
